switched template to long

+++ src/DayXX.kt
@@ -1,10 +1,10 @@
 fun main() {
-    fun part1(input: List<String>): Int {
-        return input.size
+    fun part1(input: List<String>): Long {
+        return input.size.toLong()
     }
 
-    fun part2(input: List<String>): Int {
-        return input.size
+    fun part2(input: List<String>): Long {
+        return input.size.toLong()
     }
 
     val day = 0
